Fail in-flight and queued requests with different errors
In-flight errors are failed with ConnectionClosedError, and queued (and request sent after the connection has closed) are failed with RequestNotSentError, which is a subclass of ConnectionClosedError.

The point is to make it possible to differentiate between requests that have been sent and not received a response, and requests that were never sent. The latter are always safe to retry, the former depends on the request.