Add previous response to Server#handle_error
In order for the error handler to be able to distinguish between errors
in Server#handle_request and others, this injects the resulting future
value from Server#handle_request into Server#handle_error.

To retain backwards compatibility with Server#handle_error, the dispatch
checks for a three-argument method and dispatches without the previous
response in that case. In order to support calls to super from legacy
servers, the new signature makes the previous response optional.